+++ Assets/Scripts/Pal3/Renderer/Mv3ModelRenderer.cs
@@ -441,12 +441,5 @@ namespace Pal3.Renderer
                 _meshObjects = null;
             }
         }
-        
-        public RenderMeshComponent[] GetRenderMeshComps()
-        {
-            return _renderMeshComponents;
-        }
-
-         
     }
 }
