--- Assets/Scripts/ResourceViewer/WaveFrontObjHelper.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c31b6a583a60d08469510a60efc8ebde
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
