--- Assets/Shaders/Pal3ShowNormal.shader.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 7b818b7c5700fa04b928653db303fd5f
-ShaderImporter:
-  externalObjects: {}
-  defaultTextures: []
-  nonModifiableTextures: []
-  preprocessorOverride: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
